ioNets/Blocks: fix invalid delta after a gap
If there was a gap in the datas, we would ended up computing the
delta from the previous index without taking into account that
the previous one might not be the immediate previous value but
the previous value before the gap. So fix this by skipping the
delta computation if the previous value is too old (threshold).